Move measurement code to IOpipeExecution and such.
