+++ src/main/java/com/iopipe/IOpipeService.java
@@ -331,9 +331,8 @@ public final class IOpipeService
 		boolean coldstarted = !this._coldstartflag.getAndSet(true);
 		
 		// Setup execution information
-		IOpipeMeasurement measurement = new IOpipeMeasurement(coldstarted);
 		IOpipeExecution exec = new __ActiveExecution__(this, config, __context,
-			measurement, nowtime, __input, nowmono);
+			nowtime, __input, nowmono, coldstarted);
 		
 		// Use a reference to allow the execution to be garbage collected if
 		// it is no longer referred to or is in the stack of any method.
